+++ HISTORY.rst
@@ -6,6 +6,7 @@ unreleased
 ------------------
 
 * Improve filterset class initialization during fetching export queryset
+* Make resource first filter, then order queryset
 
 1.8.0 (2025-06-30)
 ------------------