+++ HISTORY.rst
@@ -8,6 +8,7 @@ unreleased
 * Improve filterset class initialization during fetching export queryset
 * Add ability to customize the order of operation for export queryset
 * Add ability to set dataset title
+* Add field selection to the export confirmation page in the admin panel
 
 1.8.0 (2025-06-30)
 ------------------